+++ plugin/src/main/java/org/paulmanjarres/gradle/timereporter/model/PluginExtensionDefaultValues.java
@@ -2,6 +2,8 @@ package org.paulmanjarres.gradle.timereporter.model;
 
 public class PluginExtensionDefaultValues {
 
+    private PluginExtensionDefaultValues() {}
+
     public static final int longestTestsCount = 10;
     public static final int binSizeInMillis = 100;
     public static final int slowThresholdInMillis = 200;
@@ -9,5 +11,7 @@ public class PluginExtensionDefaultValues {
     public static final boolean showGroupByResult = true;
     public static final boolean showGroupByClass = true;
     public static final boolean showSlowestTests = true;
+    public static final boolean showHistogram = true;
+    public static final boolean coloredOutput = true;
     public static final boolean experimentalFeatures = false;
 }