+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.3.0
+* Colored output
+* Histogram for test time distribution
+
 ## v0.2.0
 * Support for Gradle 6.9
 * Added filter for max results on group by class